--Initial commit --Most of the source is still missing, will be added later on (so yeah, this won't compile)
This includes all the code for the CORE nullDC, whats missing are the plugins and the emitter

**** PLEASE NOTE ****
The source is old. Most of it dates back to 2006, with large parts from 2005 or 2004. There are MANY typos all over the place. There is no uniform coding style (nullDC was my first project using C/C++, and it began as a C# project)

Most of the code is written by me, parts of it are written by ZeZu, and maby others (?). Psyman was our main tester during all these years (With various others joining for short time periods) and he's slowly getting into editing the source to fix bugs ( as i no longer do it myself).

Yes, the code is ugly, hacky, buggy and often just plain wrong -- but still better than not released, righty ? 

I hope this is as much help to others as it has been to me :)

-- raz